+++ data/src/main/java/com/grusie/data/model/AlarmTimeData.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "alarm_time_table")    //앞으로 울릴 알람 시간 테이블
 data class AlarmTimeData(
     @PrimaryKey
-    val id: Int,
+    val id: Long,
     val timeInMillis: Long,
-    val alarmId: Int
+    val alarmId: Long
 )
